+++ sdk/buf.h
@@ -7,20 +7,20 @@
 typedef struct buf_t
 {
   char *data, *top;
-  size_t size, stride;
+  uint size, stride;
 
   allocator_t *allocator;
 } buf_t;
 
-buf_t *buf_create(size_t reserve, size_t stride, allocator_t *allocator);
+buf_t *buf_create(uint reserve, uint stride, allocator_t *allocator);
 void buf_delete(buf_t *buffer);
 
-int buf_push(buf_t *buffer, const void *src);
-int buf_pop(buf_t *buffer, void *dst);
+uint buf_push(buf_t *buffer, const void *src);
+uint buf_pop(buf_t *buffer, void *dst);
 void *buf_next(buf_t *buffer);
 
-int buf_full(buf_t *buffer);
-int buf_empty(buf_t *buffer);
-int bus_length(buf_t *buffer);
+uint buf_full(buf_t *buffer);
+uint buf_empty(buf_t *buffer);
+uint bus_length(buf_t *buffer);
 
 #endif
